volume cleanup lambda
