+++ alfresco-search-solr/volume_cleanup_lambda/requirements.txt
@@ -1,0 +1,7 @@
+boto3==1.26.142
+botocore==1.29.142
+jmespath==1.0.1
+python-dateutil==2.8.2
+s3transfer==0.6.1
+six==1.16.0
+urllib3==1.26.16
